feature/GH-14576: move context interfaces to root
