+++ feature-libs/cart/import-export/components/import-export/import-export-order-entries.component.spec.ts
@@ -4,8 +4,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import {
   AddOrderEntriesContext,
   GetOrderEntriesContext,
-} from '@spartacus/cart/main/components';
-import {
   OrderEntriesSource,
   OrderEntry,
   ProductData,
