+++ feature-libs/cart/import-export/components/import-export/import-export-order-entries.component.ts
@@ -1,6 +1,8 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { OrderEntriesContext } from '@spartacus/cart/main/components';
-import { ORDER_ENTRIES_CONTEXT } from '@spartacus/cart/main/root';
+import {
+  OrderEntriesContext,
+  ORDER_ENTRIES_CONTEXT,
+} from '@spartacus/cart/main/root';
 import { ContextService } from '@spartacus/storefront';
 import { Observable, of } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
