+++ feature-libs/cart/import-export/components/import-to-cart/import-entries-dialog/import-entries-dialog.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { AddOrderEntriesContext } from '@spartacus/cart/main/components';
 import {
+  AddOrderEntriesContext,
   OrderEntriesSource,
   ProductData,
   ProductImportInfo,
