+++ feature-libs/cart/main/components/cart/index.ts
@@ -10,7 +10,7 @@ export * from './cart-totals/cart-totals.component';
 export * from './cart-totals/cart-totals.module';
 export * from './mini-cart/mini-cart.component';
 export * from './mini-cart/mini-cart.module';
-export * from './order-entries-context/index';
+export * from './page-context/index';
 export * from './save-for-later/save-for-later.component';
 export * from './save-for-later/save-for-later.module';
 export * from './validation/cart-validation-components.module';
