--- feature-libs/cart/main/components/cart/order-entries-context/cart-page/index.ts
@@ -1,1 +1,0 @@
-export * from './active-cart-order-entries-context';
