--- feature-libs/cart/main/root/constants/index.ts
@@ -1,1 +1,0 @@
-export * from './order-entires.context';
