--- feature-libs/cart/main/root/constants/order-entires.context.ts
@@ -1,1 +1,0 @@
-export const ORDER_ENTRIES_CONTEXT = Symbol('ORDER_ENTRIES_CONTEXT');
