+++ feature-libs/cart/main/root/context/index.ts
@@ -1,4 +1,3 @@
 export * from './add-order-entries.context';
-export * from './cart-order-entries.context';
-export * from './cart-page/index';
 export * from './get-order-entries.context';
+export * from './order-entires.context';
