--- feature-libs/cart/quick-order/components/order-entries-context/index.ts
@@ -1,1 +1,0 @@
-export * from './quick-order-order-entries-context';
