+++ feature-libs/cart/quick-order/components/public_api.ts
@@ -1,4 +1,4 @@
 export * from './cart-quick-order-form/index';
-export * from './order-entries-context/index';
+export * from './page-context/index';
 export * from './quick-order-components.module';
 export * from './quick-order/index';
