--- feature-libs/cart/saved-cart/components/order-entries-context/index.ts
@@ -1,2 +1,0 @@
-export * from './saved-cart-details-page/saved-cart-order-entries-context';
-export * from './saved-carts-page/new-saved-cart-order-entries-context';
