+++ feature-libs/cart/saved-cart/components/page-context/saved-carts-page/new-saved-cart-order-entries.context.ts
@@ -1,10 +1,8 @@
 import { Injectable } from '@angular/core';
 import { ActionsSubject } from '@ngrx/store';
+import { CartOrderEntriesContext } from '@spartacus/cart/main/components';
 import {
   AddOrderEntriesContext,
-  CartOrderEntriesContext,
-} from '@spartacus/cart/main/components';
-import {
   Cart,
   MultiCartFacade,
   OrderEntriesSource,
