+++ feature-libs/cart/saved-cart/components/public_api.ts
@@ -1,6 +1,6 @@
 export * from './add-to-saved-cart/index';
 export * from './details/index';
 export * from './list/index';
-export * from './order-entries-context/index';
+export * from './page-context/index';
 export * from './saved-cart-components.module';
 export * from './saved-cart-form-dialog/index';
