+++ feature-libs/cart/saved-cart/components/saved-cart-components.module.ts
@@ -8,8 +8,8 @@ import { provideDefaultConfig } from '@spartacus/core';
 import { AddToSavedCartModule } from './add-to-saved-cart/add-to-saved-cart.module';
 import { SavedCartDetailsModule } from './details/saved-cart-details.module';
 import { SavedCartListModule } from './list/saved-cart-list.module';
-import { SavedCartOrderEntriesContext } from './order-entries-context/saved-cart-details-page/saved-cart-order-entries-context';
-import { NewSavedCartOrderEntriesContext } from './order-entries-context/saved-carts-page/new-saved-cart-order-entries-context';
+import { SavedCartOrderEntriesContext } from './page-context/saved-cart-details-page/saved-cart-order-entries.context';
+import { NewSavedCartOrderEntriesContext } from './page-context/saved-carts-page/new-saved-cart-order-entries.context';
 import { defaultSavedCartFormLayoutConfig } from './saved-cart-form-dialog/default-saved-cart-form-layout.config';
 import { SavedCartFormDialogModule } from './saved-cart-form-dialog/saved-cart-form-dialog.module';
