+++ feature-libs/checkout/components/order-confirmation/index.ts
@@ -1,5 +1,5 @@
 export * from './components/index';
 export * from './guards/index';
 export * from './order-confirmation.module';
-export * from './order-entries-context/index';
+export * from './page-context/index';
 export * from './replenishment-order-confirmation.module';
