--- feature-libs/checkout/components/order-confirmation/order-entries-context/index.ts
@@ -1,1 +1,0 @@
-export * from './order-confirmation-order-entries-context';
