+++ feature-libs/checkout/components/order-confirmation/page-context/index.ts
@@ -1,0 +1,1 @@
+export * from './order-confirmation-order-entries.context';
