+++ feature-libs/checkout/components/order-confirmation/page-context/order-confirmation-order-entries.context.ts
@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
-import { GetOrderEntriesContext } from '@spartacus/cart/main/components';
-import { OrderEntriesSource, OrderEntry } from '@spartacus/cart/main/root';
+import {
+  GetOrderEntriesContext,
+  OrderEntriesSource,
+  OrderEntry,
+} from '@spartacus/cart/main/root';
 import { CheckoutFacade } from '@spartacus/checkout/root';
 import { Order } from '@spartacus/order/root';
 import { Observable } from 'rxjs';
