--- feature-libs/order/components/order-entries-context/index.ts
@@ -1,1 +1,0 @@
-export * from './order-details-order-entries-context';
