+++ feature-libs/order/components/page-context/order-details-order-entries.context.ts
@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
-import { GetOrderEntriesContext } from '@spartacus/cart/main/components';
-import { OrderEntriesSource, OrderEntry } from '@spartacus/cart/main/root';
+import {
+  GetOrderEntriesContext,
+  OrderEntriesSource,
+  OrderEntry,
+} from '@spartacus/cart/main/root';
 import { Order, OrderFacade } from '@spartacus/order/root';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
