+++ feature-libs/order/components/public_api.ts
@@ -1,9 +1,9 @@
 export * from './amend-order/index';
 export * from './order-components.module';
 export * from './order-details/index';
-export * from './order-entries-context/index';
 export * from './order-history/order-history.component';
 export * from './order-history/order-history.module';
+export * from './page-context/index';
 export * from './replenishment-order-cancellation-dialog/index';
 export * from './replenishment-order-details/index';
 export * from './replenishment-order-history/index';
